Add react-router dependency, and create a separate InboxView from App
