+++ app/package.json
@@ -68,6 +68,7 @@
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-redux": "^9.2.0",
+    "react-router": "^7.6.3",
     "tailwindcss": "^4.1.11"
   }
 }
